// src/types/enums.ts
export enum EFetchStatus {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
}

export enum EGameType {
  ultra = 'ultra',
  bullet = 'bullet',
  rapid = 'rapid',
  blitz = 'blitz',
  classic = 'classic',
}

export enum EColorAttachMode {
  white = 'white',
  black = 'black',
  random = 'random',
}

export enum EPlayerColor {
  white = 'white',
  black = 'black',
}

export enum EConnectStatus {
  conected = 'conected',
  pending = 'pending',
}